+++ CHANGELOG.md
@@ -25,6 +25,9 @@
   | getWorkflowArguments("integration")
   | integration
   ```
+
+* `convert/from_10xh5_to_h5mu`: Allow reading in QC metrics, use gene ids as `.obs_names` instead of gene symbols.
+
 ## MINOR CHANGES
 
 * `dimred/umap`: Streamline UMAP parameters by adding `--obsm_output` parameter to allow choosing the output `.obsm` slot.